feat: Rename target to warehouse
