+++ internal/warehouse/doc.go
@@ -1,2 +1,2 @@
 // Package target provides the target providers implementation for the application.
-package target
+package warehouse
